AP_ESC_Telem: avoid FPE on macos
use floats for rpm in API calls

+++ libraries/AP_ESC_Telem/AP_ESC_Telem.h
@@ -89,9 +89,10 @@ public:
     bool is_telemetry_active(uint32_t servo_channel_mask) const;
 
     // callback to update the rpm in the frontend, should be called by the driver when new data is available
-    void update_rpm(const uint8_t esc_index, const uint16_t new_rpm, const float error_rate);
-
+    // can also be called from scripting
+    void update_rpm(const uint8_t esc_index, const float new_rpm, const float error_rate);
 private:
+
     // callback to update the data in the frontend, should be called by the driver when new data is available
     void update_telem_data(const uint8_t esc_index, const AP_ESC_Telem_Backend::TelemetryData& new_data, const uint16_t data_mask);
